cleanup approximation package and add reset and query
